+++ api/release-notes.md
@@ -1,2 +1,6 @@
+## v1.0.1
+- Add: `/api/driveapi/user_credentials` for getting google drive basic information
+- login endpoint now accepts more user information
+
 ## v1.0.0
 Initial Release